fix(sync): float->short UB in COB callins desynced clang-arm64 vs gcc-x86
CCobInstance::{AimWeapon,StartBuilding,SetDirection} converted radians to
TA-angle units with short(heading * RAD2TAANG). heading is in [0,2pi), so
the scaled float reaches 65535 > SHRT_MAX: the conversion is undefined
behaviour. gcc-x86 truncates through int32 and sign-extends (e.g. -16315);
clang-arm64, licensed by UB, omits the 16-bit truncation and passes the
raw int (49221) into the COB VM. The +/-65536 difference lands in COB
static-vars (e.g. BAR commanders' torsoRot via per-frame StartBuilding
while constructing) and leaks into checksummed piece rotation ~1500 frames
later -> replay/multiplayer desync between macOS and the x86 fleet.

Fix: convert explicitly via int32 (well-defined for this range), then
truncate to short, matching gcc/x86 fleet behaviour on all platforms:
short(int(heading * RAD2TAANG)).

Found by: public-replay determinism harness + full-state dump diff (single
9-ULP piece-rot divergence at f=1591) + COB opcode tracing (TURN dest
identical mod 65536 with all script inputs identical). Verified: the
previously-desyncing public demo now replays past the divergence frame
with zero sync errors on macOS-arm64.

Also: gcc build compat for the imported glibc dbl-64 (designator order via
math_config.h field order, -Wno-narrowing, __APPLE__-gated sincos_stret).

+++ rts/Sim/Units/Scripts/CobInstance.cpp
@@ -8,6 +8,7 @@
 #include "CobThread.h"
 
 #include "Game/GameHelper.h"
+#include "Sim/Misc/GlobalSynced.h"
 #include "Game/GlobalUnsynced.h"
 #include "Map/Ground.h"
 #include "Sim/Misc/GroundBlockingObjectMap.h"
@@ -237,7 +238,7 @@ void CCobInstance::WindChanged(float heading, float speed)
 {
 	ZoneScoped;
 	Call(COBFN_SetSpeed, int(speed * 3000.0f));
-	Call(COBFN_SetDirection, short(heading * RAD2TAANG));
+	Call(COBFN_SetDirection, short(int(heading * RAD2TAANG)));
 }
 
 
@@ -387,8 +388,11 @@ void CCobInstance::StartBuilding(float heading, float pitch)
 	std::array<int, 1 + MAX_COB_ARGS> callinArgs;
 
 	callinArgs[0] = 2;
-	callinArgs[1] = short(heading * RAD2TAANG);
-	callinArgs[2] = short(  pitch * RAD2TAANG);
+	// NB: float->short of out-of-range values (heading in [0,2pi) scales up to
+	// 65535 > SHRT_MAX) is UB and diverges between clang-arm64 and gcc-x86;
+	// go via int32 so the mod-2^16 truncation is explicit and identical.
+	callinArgs[1] = short(int(heading * RAD2TAANG));
+	callinArgs[2] = short(int(  pitch * RAD2TAANG));
 
 	Call(COBFN_StartBuilding, callinArgs);
 }
@@ -439,8 +443,11 @@ void CCobInstance::AimWeapon(int weaponNum, float heading, float pitch)
 	std::array<int, 1 + MAX_COB_ARGS> callinArgs;
 
 	callinArgs[0] = 2;
-	callinArgs[1] = short(heading * RAD2TAANG);
-	callinArgs[2] = short(  pitch * RAD2TAANG);
+	// NB: float->short of out-of-range values (heading in [0,2pi) scales up to
+	// 65535 > SHRT_MAX) is UB and diverges between clang-arm64 and gcc-x86;
+	// go via int32 so the mod-2^16 truncation is explicit and identical.
+	callinArgs[1] = short(int(heading * RAD2TAANG));
+	callinArgs[2] = short(int(  pitch * RAD2TAANG));
 
 	Call(COBFN_AimPrimary + COBFN_Weapon_Funcs * weaponNum, callinArgs, CBAimWeapon, weaponNum, nullptr);
 }
